:rocket: add flunt validation, and search endpoint

+++ src/TrafficTicket/TrafficTicket.Api/Controller/TrafficFineController.cs
@@ -1,10 +1,12 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using System.Net;
+using TrafficTicket.Api.DataContracts.Queries;
 using TrafficTicket.Api.Models.TrafficFine;
 using TrafficTicket.Api.Repositories;
 
 namespace TrafficTicket.Api.Controller
 {
+    [Produces("application/json")]
     [Route("api/v1/[controller]")]
     [ApiController]
     public class TrafficFineController : ControllerBase
@@ -18,15 +20,42 @@ namespace TrafficTicket.Api.Controller
 
         [HttpGet("{id}", Name = "Get")]
         [ProducesResponseType(typeof(TrafficFine), (int)HttpStatusCode.OK)]
+        [ProducesResponseType((int)HttpStatusCode.NotFound)]
+        [ProducesResponseType((int)HttpStatusCode.BadRequest)]
+        [ProducesResponseType((int)HttpStatusCode.InternalServerError)]
         public async Task<IActionResult> Get(string id)
         {
             var trafficFine = await _trafficFineRepository.GetAsycn(id);
 
-            return Ok(trafficFine ?? new TrafficFine(id));
+            if (trafficFine is null)
+            {
+                return NotFound();
+            }
+
+            return Ok(trafficFine);
+        }
+
+        [HttpGet(Name = "Search")]
+        [ProducesResponseType(typeof(TrafficFine), (int)HttpStatusCode.OK)]
+        [ProducesResponseType((int)HttpStatusCode.NotFound)]
+        [ProducesResponseType((int)HttpStatusCode.BadRequest)]
+        [ProducesResponseType((int)HttpStatusCode.InternalServerError)]
+        public async Task<IActionResult> Search([FromQuery] DateSearchQuery search)
+        {
+            var trafficFine = await _trafficFineRepository.GetByDateSearch(search);
+
+            if (trafficFine is null)
+            {
+                return NotFound();
+            }
+
+            return Ok(trafficFine);
         }
 
         [HttpPost()]
         [ProducesResponseType(typeof(TrafficFine), (int)HttpStatusCode.OK)]
+        [ProducesResponseType((int)HttpStatusCode.BadRequest)]
+        [ProducesResponseType((int)HttpStatusCode.InternalServerError)]
         public async Task<IActionResult> CreateProduct([FromBody] TrafficFine trafficFine)
         {
             await _trafficFineRepository.CreateAsync(trafficFine);
@@ -36,6 +65,8 @@ namespace TrafficTicket.Api.Controller
 
         [HttpPut]
         [ProducesResponseType(typeof(TrafficFine), (int)HttpStatusCode.OK)]
+        [ProducesResponseType((int)HttpStatusCode.BadRequest)]
+        [ProducesResponseType((int)HttpStatusCode.InternalServerError)]
         public async Task<ActionResult<TrafficFine>> Update([FromBody] TrafficFine trafficFine)
         {
             return Ok(await _trafficFineRepository.UpdateAsync(trafficFine));
@@ -43,8 +74,15 @@ namespace TrafficTicket.Api.Controller
 
         [HttpDelete("{id}", Name = "Delete")]
         [ProducesResponseType(typeof(void), (int)HttpStatusCode.OK)]
+        [ProducesResponseType((int)HttpStatusCode.BadRequest)]
+        [ProducesResponseType((int)HttpStatusCode.InternalServerError)]
         public async Task<IActionResult> Delete(string id)
         {
+            if (string.IsNullOrEmpty(id))
+            {
+                return BadRequest();
+            }
+
             await _trafficFineRepository.DeleteAsync(id);
             return NoContent();
         }
